requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2018.1.18
chardet==3.0.4
idna==2.6
lxml==4.2.0
PyYAML==3.12
redis==2.10.6
requests==2.18.4
simplejson==3.13.2
tornado==5.0
urllib3==1.22
